Extract build-status assembly out of submissions.ts
Phase 3 Wave B, batch 3: pulls the closure cluster that assembles a
submission status response's channel events, media, playables, and
prior-round history — loadBuildEvents, loadBuildPreviews, loadBuildShots,
buildMedia, buildPlayables, localizeRevisions, localizeEvents,
loadPriorRounds, and attachBuildEvents — into
apps/api/src/delivery/build-status.ts as createBuildStatusAssembler(),
a plain factory (no Fastify routes of its own, since its one caller
route stays in submissions.ts alongside the still-central refreshStatus).

revisionOriginOf moves with it (only used inside loadPriorRounds, plus
one remaining use in submissions.ts's nativeJobStatus, which now imports
it back). The three call sites that invalidated the events cache directly
(invalidateDeliveryCaches, the staged-preview onPublished callback, and
the agent-channel onEvent callback) now go through the new module's
exported invalidateEvents(issueNumber) instead of reaching into a cache
that no longer lives in this file.

submissions.ts drops from 5529 to 5197 lines with no behavior change.
